% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_analyticTs.R
\name{calc_analyticTs}
\alias{calc_analyticTs}
\title{Applies the analytical formula to calculate temperature time series at one depth}
\usage{
calc_analyticTs(param, z, nbDays = 3)
}
\arguments{
\item{param}{the vector of parameters}

\item{z}{(in m) a vector of depths at which to calculate the temperature}

\item{nbDays}{(in days) length of output time series}

\item{Deltat}{(in s) time resolution of output time series}
}
\value{
temperature time series at depth z
}
\description{
Applies the analytical formula to calculate temperature time series at one depth
}

