% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_analyticTs.R
\name{calc_analyticTs_timeDepth}
\alias{calc_analyticTs_timeDepth}
\title{Applies the analytical formula to calculate temperature at one depth, time combination}
\usage{
calc_analyticTs_timeDepth(param, t, z)
}
\arguments{
\item{param}{the vector of parameters}

\item{t}{(in s) time at which to calculate the temperature}

\item{z}{(in m) a vector of depths at which to calculate the temperature
Depths need to be between 0 and maximum depth.}
}
\value{
temperature at depth z and time t for parameters param
}
\description{
Applies the analytical formula to calculate temperature at one depth, time combination
}

