% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_likelihood.R
\name{calc_center}
\alias{calc_center}
\title{Calculates the mse between measurements and simulations}
\usage{
calc_center(df_all)
}
\arguments{
\item{df_all}{the dataframe containing measurements (in meas) and simulations
(in columns starting with sim)}
}
\value{
a dataframe where the mean of each timeseries has been extracted
  (grouped by depth and simulation/measurement)
}
\description{
Calculates the mse between measurements and simulations
}

