% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calc_comb_phy}
\alias{calc_comb_phy}
\title{Calculate combinations of physical parameters from ranges}
\usage{
calc_comb_phy(param, nb_lambda_s = 10, nb_permeability = 10,
  nb_rho_s = 10, nb_c_s = 10, nb_n = 10)
}
\arguments{
\item{param}{the vector of parameters}

\item{nb_lambda_s}{the number of lambda_s values}

\item{nb_permeability}{the number of permeability values}

\item{nb_rho_s}{the number of rho_s values}

\item{nb_c_s}{the number of c_s values}

\item{nb_n}{the number of n values}
}
\value{
a dataframe of combinations of parameters
}
\description{
Calculate combinations of physical parameters from ranges
}

