% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calc_eq_fromPhyDf}
\alias{calc_eq_fromPhyDf}
\title{Add equivalent parameters to dataframe of physical parameters}
\usage{
calc_eq_fromPhyDf(param, df_phys)
}
\arguments{
\item{param}{the vector of parameters}

\item{df_phys}{the dataframe containing physical parameter values}
}
\value{
the dataframe of physical values with fields alpha_e and kappa_e
}
\description{
Add equivalent parameters to dataframe of physical parameters
}

