% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calc_inv}
\alias{calc_inv}
\title{Inverse operation: Update both lambda_s and permeability from param}
\usage{
calc_inv(param)
}
\arguments{
\item{the}{vector of parameters}
}
\value{
the vector of parameters with updated lambda_s and permeability values
}
\description{
Inverse operation: Update both lambda_s and permeability from param
}

