% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calc_inv_lambda_s}
\alias{calc_inv_lambda_s}
\title{Inverse operation: calculate lambda_s from vector of parameters}
\usage{
calc_inv_lambda_s(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
param with updated lambda_s field
}
\description{
Inverse operation: calculate lambda_s from vector of parameters
}

