% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calc_inv_permeability}
\alias{calc_inv_permeability}
\title{Inverse operation: calculate permeability from vector of parameters}
\usage{
calc_inv_permeability(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
param with updated permeability field
}
\description{
Inverse operation: calculate permeability from vector of parameters
}

