% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calc_lambda_m}
\alias{calc_lambda_m}
\title{Calculate equivalent thermal conductivity from vector of parameters according
to the wood formula}
\usage{
calc_lambda_m(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
the same vector of parameters param with updated lambda_m field
}
\description{
Calculate equivalent thermal conductivity from vector of parameters according
to the wood formula
}

