% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_likelihood.R
\name{calc_like}
\alias{calc_like}
\title{Calculates the likelihood}
\usage{
calc_like(df_eq_comb, df_all, param, date_min = NULL, center = F)
}
\arguments{
\item{df_eq_comb}{the dataframe containing combinations of
alpha_e, kappa_e and sim_idx}

\item{df_all}{the dataframe containing measurements (in meas)
and simulations (in columns starting with sim).
sim_idx is the key linking
a set of reduced parameters in df_eq_comb
to corresponding columns of df_all
check that the names of columns in df_all
are the same than contents of the sim_idx field.}

\item{param}{the vector parameters}

\item{date_min}{date specifying start of rmse calculation, before this date
values will be discarded (given as POSIXct object)}

\item{center}{boolean indicating whether to center timeseries before
calculating the rmse}
}
\value{
the dataframe df_eq_comb with corresponding likelihood
(in like). Also returns mse and logLike.
}
\description{
Calculates the likelihood
}

