% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calc_Peclet}
\alias{calc_Peclet}
\title{Calculate Peclet number from vector of parameters}
\usage{
calc_Peclet(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
the Peclet number
}
\description{
Calculate Peclet number from vector of parameters
}

