% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calc_q}
\alias{calc_q}
\title{Calculate flux from vector of parameters}
\usage{
calc_q(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
the flow value (in m/s)
}
\description{
Calculate flux from vector of parameters
}

