% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calc_range_eq}
\alias{calc_range_eq}
\title{Calculate range of reduced parameters from vector of parameters}
\usage{
calc_range_eq(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
the range of reduced parameters
}
\description{
Calculate range of reduced parameters from vector of parameters
}

