% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calc_rho_m_c_m}
\alias{calc_rho_m_c_m}
\title{Calculate rho_m * c_m from vector of parameters}
\usage{
calc_rho_m_c_m(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
the same vector of parameters param with updated lambda_m field
}
\description{
Calculate rho_m * c_m from vector of parameters
}

