% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calculate_eq}
\alias{calculate_eq}
\title{Calculate equivalent reduced parameters from vector of parameters}
\usage{
calculate_eq(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
the equivalent reduced parameters
}
\description{
Calculate equivalent reduced parameters from vector of parameters
}

