% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calculate_lambda_m}
\alias{calculate_lambda_m}
\title{Calculate equivalent thermal conductivity from vector of parameters according to the wood formula}
\usage{
calculate_lambda_m(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
the equivelent thermal conductivity
}
\description{
Calculate equivalent thermal conductivity from vector of parameters according to the wood formula
}

