% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calculatePeclet}
\alias{calculatePeclet}
\title{Calculate Peclet number from vector of parameters}
\usage{
calculatePeclet(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
the Peclet number
}
\description{
Calculate Peclet number from vector of parameters
}

