% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calculatePeclet_eq}
\alias{calculatePeclet_eq}
\title{Calculate Peclet number from vector of equivalent parameters}
\usage{
calculatePeclet_eq(param_eq, param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
the Peclet number
}
\description{
Calculate Peclet number from vector of equivalent parameters
}

