% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_BC}
\alias{check_BC}
\title{Check the consistency in boundary conditions
between 2 vector of parameters}
\usage{
check_BC(param_1, param_2)
}
\arguments{
\item{param_1}{the first vector of parameters}

\item{param_2}{the second vector of parameters}
}
\value{
boolean for whether boundary conditions are the same
}
\description{
Check the consistency in boundary conditions
between 2 vector of parameters
}

