% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{check_consistency}
\alias{check_consistency}
\title{Check consistency of physical and reduced parameters in param}
\usage{
check_consistency(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
boolean indicating whether
the definition of equivalent parameters is consistent
}
\description{
Check consistency of physical and reduced parameters in param
}

