% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_isInBounds}
\alias{check_isInBounds}
\title{Check that physical parameter values are within bounds}
\usage{
check_isInBounds(param)
}
\arguments{
\item{param}{the first vector of parameters}
}
\value{
boolean for whether boundary conditions are the respected
}
\description{
Check that physical parameter values are within bounds
}

