% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{csv2param}
\alias{csv2param}
\title{Imports vector of parameter from csv file}
\usage{
csv2param(csv_file)
}
\arguments{
\item{csv_file}{path to the csv file}
}
\value{
the vector of parameters
}
\description{
Imports vector of parameter from csv file
}

