% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_comb.R
\name{gen_atCenter}
\alias{gen_atCenter}
\title{Generates sequence of values at center of intervals}
\usage{
gen_atCenter(n, val_min, val_max, isLog10 = F)
}
\arguments{
\item{n}{length of the desired vector}

\item{val_min}{minimum value}

\item{val_max}{maximum value}

\item{isLog10}{generates on the log10 scale}
}
\value{
the sequence of values
}
\description{
Generates sequence of values at center of intervals
}

