% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_comb.R
\name{gen_comb_phy_atCenter}
\alias{gen_comb_phy_atCenter}
\title{Generates combinations of physical parameters from ranges
at center of grids}
\usage{
gen_comb_phy_atCenter(param, nb_lambda_s = 5, nb_permeability = 10,
  nb_rho_s = 2, nb_c_s = 2, nb_n = 2)
}
\arguments{
\item{param}{the vector of parameters}

\item{nb_lambda_s}{the number of lambda_s values}

\item{nb_permeability}{the number of permeability values}

\item{nb_rho_s}{the number of rho_s values}

\item{nb_c_s}{the number of c_s values}

\item{nb_n}{the number of n values}
}
\value{
a dataframe of combinations of parameters
}
\description{
Generates combinations of physical parameters from ranges
at center of grids
}

