% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{get_level_z_fromParam}
\alias{get_level_z_fromParam}
\title{Calculate table of correspondance between z and z_idx
from the vector of parameters}
\usage{
get_level_z_fromParam(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
the table of correspondance
}
\description{
Calculate table of correspondance between z and z_idx
from the vector of parameters
}

