% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_BC}
\alias{names_BC}
\title{names of parameters for boundary conditions}
\format{a vector of characters}
\source{
self-defined
}
\usage{
names_BC
}
\description{
A vector containing the names of boundary conditions parameters
}
\keyword{datasets}

