% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_phy}
\alias{names_phy}
\title{names of physical parameters}
\format{a vector of characters}
\source{
self-defined
}
\usage{
names_phy
}
\description{
A vector containing the names of physical parameters of interest
}
\keyword{datasets}

