% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_red}
\alias{names_red}
\title{names of reduced set of parameters}
\format{a vector of characters}
\source{
self-defined
}
\usage{
names_red
}
\description{
A vector containing the names of reduced parameters of interest
}
\keyword{datasets}

