% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_pdf.R
\name{pdf2cdf}
\alias{pdf2cdf}
\title{Calculates cdf from pdf}
\usage{
pdf2cdf(df_pdf)
}
\arguments{
\item{df_pdf}{the dataframe containing the univariate pdf}
}
\value{
a dataframe containing x and cdf values
}
\description{
Calculates cdf from pdf
}

