% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_pdf.R
\name{pdf_define_unif}
\alias{pdf_define_unif}
\title{Defines uniform distribution between 2 bounds}
\usage{
pdf_define_unif(x_range, n = 50)
}
\arguments{
\item{x_range}{a numerical vector of 2 elements
containing the minimal and maximal value of RV}

\item{n}{number of points at which to evaluate the density}
}
\value{
a dataframe correponding to the pdf of the RV
}
\description{
Defines uniform distribution between 2 bounds
}

