% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_pdf.R
\name{pdf_marginalize}
\alias{pdf_marginalize}
\title{Marginalizes a multivariate pdf}
\usage{
pdf_marginalize(df_pdf_mv, colName_var, colName_pdf, normalize = T)
}
\arguments{
\item{df_pdf_mv}{the dataframe containing the multivariate pdf, with variable
values and pdf values}

\item{colName_var}{the variable for which to calculate the marginalized pdf}

\item{colName_pdf}{the column containing the actual pdf value}

\item{normalize}{boolean indicating whether to normalize the marginal pdf}
}
\value{
a dataframe correponding to the marginalized pdf
}
\description{
Marginalizes a multivariate pdf
}

