% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_pdf.R
\name{pdf_normalize}
\alias{pdf_normalize}
\title{Normalizes a pdf object - not exported}
\usage{
pdf_normalize(df_pdf)
}
\arguments{
\item{df_pdf}{a dataframe containing columns x and p_x}
}
\value{
a dataframe correponding to the normalized pdf of the RV
}
\description{
Normalizes a pdf object - not exported
}

