% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_pdf.R
\name{pdf_postInterval}
\alias{pdf_postInterval}
\title{Calculates posterior interval estimate from pdf}
\usage{
pdf_postInterval(df_pdf, alpha = 0.05)
}
\arguments{
\item{df_pdf}{the dataframe containing the univariate pdf}

\item{alpha}{a number between 0 and 0.5}
}
\value{
a vector containing 2 numbers corresponding to
}
\description{
Calculates posterior interval estimate from pdf
}

