% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hz1d.R
\name{plot_head_differential_fromMeas}
\alias{plot_head_differential_fromMeas}
\title{Plots head differential measurement timeseries from dataframe from csv file}
\usage{
plot_head_differential_fromMeas(df_measGinette, dates_range = NULL)
}
\arguments{
\item{df_measGinette}{the dataframe of measurements
(as outputted by read_measGinette)}

\item{dates_range}{range of dates for x-axis}
}
\value{
a ggplot object
}
\description{
Plots head differential measurement timeseries from dataframe from csv file
}

