% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hz1d.R
\name{plot_hz1d_t}
\alias{plot_hz1d_t}
\title{Plots temperature measurement timeseries}
\usage{
plot_hz1d_t(param, df_hz1d_t)
}
\arguments{
\item{param}{the vector of parameters}

\item{df_hz1d_t}{the dataframe of measurements}
}
\value{
a ggplot object
}
\description{
Plots temperature measurement timeseries
}

