% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_simusGinette.R
\name{read_measGinette}
\alias{read_measGinette}
\title{Reads measurements given to Ginette (both boundary conditions and
intermediary temperatures)}
\usage{
read_measGinette(pathCsv)
}
\arguments{
\item{pathCsv}{path to csv containing files (typically in
field_database/1_data-hz_to_ginette/data_hz/data/03_manual_corrections/)}
}
\value{
a dataframe in long format containing cleaned values of measurements
  on the field
}
\description{
Reads measurements given to Ginette (both boundary conditions and
intermediary temperatures)
}

