% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_simusGinette.R
\name{read_parameterSets}
\alias{read_parameterSets}
\title{Reads combinations of physical parameters from list of files of parameters
given to Ginette}
\usage{
read_parameterSets(pathFold)
}
\arguments{
\item{pathFold}{path to folder where}
}
\value{
a dataframe containing values of physical parameters (named as in
  names_phy) for each combination of iFile and iLine
}
\description{
Reads combinations of physical parameters from list of files of parameters
given to Ginette
}

