% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_analyticTs.R
\name{sim_analytic_fromRed}
\alias{sim_analytic_fromRed}
\title{Applies the analytical formula to calculate temperature time series at one depth}
\usage{
sim_analytic_fromRed(param, df_red)
}
\arguments{
\item{param}{the vector of parameters}

\item{df_red}{the dataframe containing combinations of
reduced parameters alpha_e and kappa_e for which to calculate the analytical solution}
}
\value{
a dataframe containing temperature timeseries
corresponding to reduced parameters
}
\description{
Applies the analytical formula to calculate temperature time series at one depth
}

