% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{update_all_eq}
\alias{update_all_eq}
\title{Updates the definition of lambda_m, rho_m_c_m and reduced parameters}
\usage{
update_all_eq(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
param the updated vector of parameters
}
\description{
Updates the definition of lambda_m, rho_m_c_m and reduced parameters
}

