% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_params.R
\name{calc_eq}
\alias{calc_eq}
\title{Calculate reduced parameters from physical parameters}
\usage{
calc_eq(param)
}
\arguments{
\item{param}{the vector of parameters}
}
\value{
the same vector of parameters param with updated rho_m_c_m field
}
\description{
Calculate reduced parameters from physical parameters
}

