% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_likelihood.R
\name{calc_rmse}
\alias{calc_rmse}
\title{Calculates the rmse
between measurements and simulations}
\usage{
calc_rmse(df_all)
}
\arguments{
\item{df_all}{the dataframe containing measurements (in meas)
and simulations (in columns starting with sim)}
}
\value{
a dataframe containing rmse values for each simulation
}
\description{
Calculates the rmse
between measurements and simulations
}

