% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyticTs.R
\name{calculateSyntheticTs}
\alias{calculateSyntheticTs}
\title{Applies the analytical formula to calculate temperature time series at one depth}
\usage{
calculateSyntheticTs(param, z, nbDays = 3, Deltat = 15 * 60)
}
\arguments{
\item{param}{the vector of parameters}

\item{z}{depth at which to calculate the temperature}

\item{nbDays}{length of output time series, in days}

\item{Deltat}{time resolution of output time series}
}
\value{
temperature time series at depth z
}
\description{
Applies the analytical formula to calculate temperature time series at one depth
}

