% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hz1d.R
\name{plot_hz1d_t}
\alias{plot_hz1d_t}
\title{Plots temperature measurement timeseries}
\usage{
plot_hz1d_t(param, df_hz1d_t)
}
\arguments{
\item{param}{the vector of parameters}

\item{df_hz1d_t}{the dataframe of measurements}

\item{z}{(in m) a vector of depths at which to calculate the temperature
Depths need to be between 0 and maximum depth.}
}
\value{
temperature at depth z and time t for parameters param
}
\description{
Plots temperature measurement timeseries
}

